// src/Enums/FiedTypes.ts
enum FieldTypes {
    TEXT = "text",
    EMAIL = "email",
    FILE = "file",
    SELECT = "select",
    REFERENCE = "reference",
    PASSWORD = "password",
    IMAGE = "image",
    TEXTAREA = "textarea",
    DATE = "date",
    DATETIME = "date-time",
    SCRIPT = "script",
    TIMESTAMP = "timestamp",
    NUMBER = "number",
    SELECTORADD = "select_or_add",
    MULTISELECT = "multi_select"
}

export default FieldTypes;